Try to ensure we don't throw away the PIN unnecessarily
Now on an engine error we decode the error value and determine if the issue is
due to a true PIN error or not. If it is delete the PIN as usual, but if it
isn't let the PIN be.

This does add a build-time dependency to libp11 in order to determine the
OpenSSL lib identifier from GET_ERR_LIB.

Signed-off-by: Mike Gerow <gerow@google.com>